integration test fix

+++ plugins/modules/win_dns_record.ps1
@@ -96,41 +96,57 @@ if ($null -ne $records) {
     foreach ($value in $values) {
         $required_values[$value.ToString()] = $null
     }
-
     foreach ($record in $records) {
-        $record_value = $record.RecordData.$record_argument_name.ToString()
+      $record_value = $record.RecordData.$record_argument_name.ToString()
+      if($type -eq 'SRV'){
         $record_port_old = $record.RecordData.Port.ToString()
         $record_priority_old = $record.RecordData.Priority.ToString()
         $record_weight_old = $record.RecordData.Weight.ToString()
-        
-        if ($required_values.ContainsKey($record_value)) {
+          if ($required_values.ContainsKey($record_value)) {
             # This record matches one of the values; but does it match the TTL?
-            if ($record.TimeToLive -ne $ttl) {
-                $new_record = $record.Clone()
-                $new_record.TimeToLive = $ttl
-                $new_record.RecordData.Port = $port
-                $new_record.RecordData.Priority = $priority
-                $new_record.RecordData.Weight = $weight
-                Set-DnsServerResourceRecord -ZoneName $zone -OldInputObject $record -NewInputObject $new_record -WhatIf:$module.CheckMode @extra_args
+              if ($record.TimeToLive -ne $ttl -or $port -ne $record_port_old -or $priority -ne $record_priority_old -or $weight -ne $record_weight_old) {
+                    $new_record = $record.Clone()
+                    $new_record.TimeToLive = $ttl
+                    $new_record.RecordData.Port = $port
+                    $new_record.RecordData.Priority = $priority
+                    $new_record.RecordData.Weight = $weight
+                    Set-DnsServerResourceRecord -ZoneName $zone -OldInputObject $record -NewInputObject $new_record -WhatIf:$module.CheckMode @extra_args
 
-                $changes.before += "[$zone] $($record.HostName) $($record.TimeToLive.TotalSeconds) IN $type $record_value $record_port_old $record_weight_old $record_priority_old`n"
-                $changes.after += "[$zone] $($record.HostName) $($ttl.TotalSeconds) IN $type $record_value $port $weight $priority`n"
+                    $changes.before += "[$zone] $($record.HostName) $($record.TimeToLive.TotalSeconds) IN $type $record_value $record_port_old $record_weight_old $record_priority_old`n"
+                    $changes.after += "[$zone] $($record.HostName) $($ttl.TotalSeconds) IN $type $record_value $port $weight $priority`n"
+                    $module.Result.changed = $true
+              }
+              # Cross this one off the list, so we don't try adding it later
+              $required_values.Remove($record_value)
+          } else {
+                # This record doesn't match any of the values, and must be removed
+                $record | Remove-DnsServerResourceRecord -ZoneName $zone -Force -WhatIf:$module.CheckMode @extra_args
+                $changes.before += "[$zone] $($record.HostName) $($record.TimeToLive.TotalSeconds) IN $type $record_value`n"
                 $module.Result.changed = $true
             }
+            # Whatever is left in $required_values needs to be added
+            $values = $required_values.Keys
+      }
+        else{
+            if ($required_values.ContainsKey($record_value)) {
+                if ($record.TimeToLive -ne $ttl) {
+                    $new_record = $record.Clone()
+                    $new_record.TimeToLive = $ttl
+                    Set-DnsServerResourceRecord -ZoneName $zone -OldInputObject $record -NewInputObject $new_record -WhatIf:$module.CheckMode @extra_args
+                    $changes.before += "[$zone] $($record.HostName) $($record.TimeToLive.TotalSeconds) IN $type $record_value`n"
+                    $changes.after += "[$zone] $($record.HostName) $($ttl.TotalSeconds) IN $type $record_value`n"
+                    $module.Result.changed = $true
+                }
+                $required_values.Remove($record_value)
+            } else {
+                $record | Remove-DnsServerResourceRecord -ZoneName $zone -Force -WhatIf:$module.CheckMode @extra_args
 
-            # Cross this one off the list, so we don't try adding it later
-            $required_values.Remove($record_value)
-        } else {
-            # This record doesn't match any of the values, and must be removed
-            $record | Remove-DnsServerResourceRecord -ZoneName $zone -Force -WhatIf:$module.CheckMode @extra_args
-
-            $changes.before += "[$zone] $($record.HostName) $($record.TimeToLive.TotalSeconds) IN $type $record_value`n"
-            $module.Result.changed = $true
+                $changes.before += "[$zone] $($record.HostName) $($record.TimeToLive.TotalSeconds) IN $type $record_value`n"
+                $module.Result.changed = $true
+            }
+            $values = $required_values.Keys
         }
     }
-
-    # Whatever is left in $required_values needs to be added
-    $values = $required_values.Keys
 }
